+++ sample-apps/react-native/dogfood/src/components/AudioRoom/LiveButtons.tsx
@@ -3,9 +3,8 @@ import {
   OwnCapability,
   Restricted,
   useCall,
-  useCallCallingState,
+  useCallStateHooks,
   useI18n,
-  useIsCallLive,
 } from '@stream-io/video-react-native-sdk';
 import React from 'react';
 import { Button } from 'react-native';
@@ -13,6 +12,7 @@ import { Button } from 'react-native';
 export default function LiveButtons({ onJoined }: { onJoined: () => void }) {
   // this utility hook returns the call object from the <StreamCall /> context
   const call = useCall();
+  const { useIsCallLive, useCallCallingState } = useCallStateHooks();
   // this utility hook is a wrapper around the `call.state.metadata$` observable,
   // and it will emit a new value whenever the call goes live or stops being live.
   // we can use it to update the button text or adjust any other UI elements